Restart of the git repo for this blog.
We had a git history already. We used the `develop` branch to as the main development branch.
We used the `master` branch to build and publish this site.

Today, we start working on an iteration on this website.
Thats why we want to start clean and make it more easy.